Introduce the krun_set_data_disk API.
This API adds the possibility to introduce a second block device,
to a TEE.

It is assumed that the root disk contains a symmetric key (secret),
and code to encrypt/decrypt the data disk before use.
The recommended setup is to included cryptsetup on the root disk
and use that to safely access the data disk.